+++ packages/web-components/fast-foundation/src/tree-item/tree-item.ts
@@ -5,7 +5,6 @@ import {
     SyntheticViewTemplate,
 } from "@microsoft/fast-element";
 import { isHTMLElement } from "@microsoft/fast-web-utilities";
-import { ARIAGlobalStatesAndProperties } from "../patterns/aria-global.js";
 import { StartEnd, StartEndOptions } from "../patterns/start-end.js";
 import { applyMixins } from "../utilities/apply-mixins.js";
 
@@ -58,12 +57,6 @@ export class FASTTreeItem extends FASTElement {
     public expanded: boolean = false;
     protected expandedChanged(prev: boolean | undefined, next: boolean): void {
         if (this.$fastController.isConnected) {
-            this.ariaExpanded = this.childItems?.length
-                ? next
-                    ? "true"
-                    : "false"
-                : null;
-
             this.$emit("expanded-change", this);
         }
     }
@@ -78,7 +71,6 @@ export class FASTTreeItem extends FASTElement {
     public selected: boolean;
     protected selectedChanged(prev: boolean | undefined, next: boolean): void {
         if (this.$fastController.isConnected) {
-            this.ariaSelected = next ? "true" : "false";
             this.$emit("selected-change", this);
         }
     }
@@ -91,11 +83,6 @@ export class FASTTreeItem extends FASTElement {
      */
     @attr({ mode: "boolean" })
     public disabled: boolean;
-    protected disabledChanged(prev: boolean | undefined, next: boolean): void {
-        if (this.$fastController.isConnected) {
-            this.ariaDisabled = next ? "true" : "false";
-        }
-    }
 
     /**
      *  Reference to the expand/collapse button
@@ -213,54 +200,11 @@ export class FASTTreeItem extends FASTElement {
 }
 
 /**
- * Includes ARIA states and properties relating to the ARIA textbox role
- *
- * @public
- */
-export class DelegatesARIATreeItem {
-    /**
-     * See {@link https://www.w3.org/TR/wai-aria-1.2/#treeitem} for more information
-     * @public
-     * @remarks
-     * HTML Attribute: `aria-disabled`
-     */
-    @observable
-    public ariaDisabled: "true" | "false" | string | null;
-
-    /**
-     * See {@link https://www.w3.org/TR/wai-aria-1.2/#treeitem} for more information
-     * @public
-     * @remarks
-     * HTML Attribute: `aria-expanded`
-     */
-    @observable
-    public ariaExpanded: "true" | "false" | string | null;
-
-    /**
-     * See {@link https://www.w3.org/TR/wai-aria-1.2/#treeitem} for more information
-     * @public
-     * @remarks
-     * HTML Attribute: `aria-expanded`
-     */
-    @observable
-    public ariaSelected: "true" | "false" | string | null;
-}
-
-/**
- * Mark internal because exporting class and interface of the same name
- * confuses API documenter.
- * TODO: https://github.com/microsoft/fast/issues/3317
- * @internal
- */
-export interface DelegatesARIATreeItem extends ARIAGlobalStatesAndProperties {}
-applyMixins(DelegatesARIATreeItem, ARIAGlobalStatesAndProperties);
-
-/**
  * Mark internal because exporting class and interface of the same name
  * confuses API documenter.
  * TODO: https://github.com/microsoft/fast-dna/issues/3317
  * @internal
  */
 /* eslint-disable-next-line */
-export interface FASTTreeItem extends StartEnd, DelegatesARIATreeItem {}
-applyMixins(FASTTreeItem, StartEnd, DelegatesARIATreeItem);
+export interface FASTTreeItem extends StartEnd {}
+applyMixins(FASTTreeItem, StartEnd);