+++ packages/web-components/fast-element/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @microsoft/fast-element
 
-This log was last generated on Mon, 03 Oct 2022 23:44:38 GMT and should not be manually modified.
+This log was last generated on Mon, 10 Oct 2022 20:28:02 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 2.0.0-beta.13
+
+Mon, 10 Oct 2022 20:28:02 GMT
+
+### Changes
+
+- fix: prevent event callbacks when controller is not bound (roeisenb@microsoft.com)
+
+## 2.0.0-beta.12
+
+Thu, 06 Oct 2022 23:21:20 GMT
+
+### Changes
+
+- fix: repeat update view context positioning (roeisenb@microsoft.com)
+
+## 2.0.0-beta.11
+
+Wed, 05 Oct 2022 23:26:01 GMT
+
+### Changes
+
+- fix: removed a shadowRoot leak from ElementController (nicholasrice@users.noreply.github.com)
 
 ## 2.0.0-beta.10